requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
paddleocr==2.9.1
paddlepaddle==3.0.0
Pillow==10.1.0
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2
tokenizers==0.15.0